Agregando configuracion de seilog para peticiones http, por medio del…

+++ src/SG.Infrastructure.Auth/JwtAuthentication/JwtOptions.cs
@@ -6,5 +6,5 @@ public sealed class JwtOptions
     public string Issuer { get; init; }  = default!; 
     public string Audience { get; set; }  = default!; 
     public string SigningKey { get; set; }  = default!; 
-    public int ExpiratioMinutes { get; set; }
+    public double ExpiratioMinutes { get; set; }
 }
